ootd api 완성

+++ WSIWT_BACK/src/main/java/com/example/wsiwt_back/service/OOTDService.java
@@ -2,6 +2,7 @@ package com.example.wsiwt_back.service;
 
 import com.example.wsiwt_back.domain.ootd.OOTD;
 import com.example.wsiwt_back.domain.ootd.OOTDRepository;
+import com.example.wsiwt_back.domain.user.UserRepository;
 import com.example.wsiwt_back.web.dto.ootd.OOTDSaveRequestDto;
 import com.example.wsiwt_back.web.dto.ootd.OOTDUpdateRequestDto;
 import lombok.RequiredArgsConstructor;
@@ -18,7 +19,7 @@ import java.util.List;
 public class OOTDService {
 
     private final OOTDRepository ootdRepository;
-
+    private final UserRepository userRepository;
 
     public Long save(OOTD entity){
         //String url = "images/winter_hat.png";
@@ -33,14 +34,30 @@ public class OOTDService {
         return ootdRepository.findAll();
     }
 
-    public void delete(Long id){
+    public void delete(Long id,String userId){
+
+        OOTD ootd = ootdRepository.findById(id).orElseThrow(() -> new IllegalArgumentException("해당 게시글이 없습니다. id="+id));
+        System.out.println(userId);
+        System.out.println(ootd.getUserId());
+        if(!ootd.getUserId().equals(userId)  ){
+            log.warn("글을 작성한 사람만 삭제할 수 있습니다.");
+            throw new RuntimeException("글을 작성한 사람만 삭제할 수 있습니다.");
+        }
+
         ootdRepository.deleteById(id);
     }
 
     @Transactional
-    public OOTD update(Long id, OOTDUpdateRequestDto requestDto){
+    public OOTD update(Long id,String userId, OOTDUpdateRequestDto requestDto){
+
+
         OOTD ootd = ootdRepository.findById(id).orElseThrow(() -> new IllegalArgumentException("해당 게시글이 없습니다. id="+id));
-        ootd.update(requestDto.getContent());
+        if(!ootd.getUserId().equals(userId)  ){
+            log.warn("글을 작성한 사람만 수정할 수 있습니다.");
+            throw new RuntimeException("글을 작성한 사람만 수정할 수 있습니다.");
+        }
+
+        ootd.update(requestDto.getContent(), requestDto.getUrl());
         return ootd;
     }
